LayoutTests/media/video-main-content-deny-obscured-expected.txt


RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForRateChange,OverrideUserGestureRequirementForMainContent"))
EVENT(canplaythrough)
Did not begin playing OK
END OF TEST
